追加：メール送信機能を追加、アカウントの新規登録フォームの実装（5割完成）、テーブル・カラムの追加

+++ app/Http/Controllers/ProfileController.php
@@ -7,6 +7,7 @@ use Illuminate\Contracts\Auth\MustVerifyEmail;
 use Illuminate\Http\RedirectResponse;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Hash;
 use Illuminate\Support\Facades\Redirect;
 use Inertia\Inertia;
 use Inertia\Response;
@@ -14,7 +15,9 @@ use Inertia\Response;
 class ProfileController extends Controller
 {
     /**
-     * Display the user's profile form.
+     * アカウント設定画面を表示する
+     * @param Request $request
+     * @return Response
      */
     public function edit(Request $request): Response
     {
@@ -25,16 +28,15 @@ class ProfileController extends Controller
     }
 
     /**
-     * Update the user's profile information.
+     * アカウント情報を更新する
+     * @param ProfileUpdateRequest $request
+     * @return RedirectResponse
      */
     public function update(ProfileUpdateRequest $request): RedirectResponse
     {
-        $request->user()->fill($request->validated());
-
-        if ($request->user()->isDirty('email')) {
-            $request->user()->email_verified_at = null;
+        if ($request->user()->isDirty('u_email')) {
+            $request->user()->u_email_verified_at = null;
         }
-
         $request->user()->save();
 
         return Redirect::route('profile.edit');
@@ -45,19 +47,22 @@ class ProfileController extends Controller
      */
     public function destroy(Request $request): RedirectResponse
     {
-        $request->validate([
-            'password' => ['required', 'current_password'],
-        ]);
+        if (Auth::check()) {
+//            $user = Auth::user();
+//            var_dump(Hash::check($request->u_password, $request->user()->u_password));
+//
+//            $request->validate([
+//                'u_password' => ['required', 'current_password'],
+//            ]);
+            $user = $request->user();
 
-        $user = $request->user();
+            Auth::logout();
+            $user->delete();
+            $request->session()->invalidate();
+            $request->session()->regenerateToken();
 
-        Auth::logout();
-
-        $user->delete();
-
-        $request->session()->invalidate();
-        $request->session()->regenerateToken();
-
-        return Redirect::to('/');
+            return Redirect::to('/');
+        }
+        return Redirect::to('/login');
     }
 }
